+++ DTCD-StorageSystem/src/Plugin.js
@@ -1,5 +1,6 @@
 import { SystemPlugin, LogSystemAdapter } from 'SDK';
 import { SessionModule } from './modules/SessionModule';
+import { TokenModule } from './modules/TokenModule';
 import pluginMeta from './Plugin.Meta';
 
 /**
@@ -13,6 +14,12 @@ export class StorageSystem extends SystemPlugin {
    * @property @private
    */
   #sessionModule;
+
+  /**
+   * Private instance of the TokenModule class.
+   * @property @private
+   */
+  #tokenModule;
 
   /**
    * Private instance of the LogSystemAdapter class.
@@ -33,6 +40,7 @@ export class StorageSystem extends SystemPlugin {
     this.#logSystem.debug(`Start of ${systemName} creation`);
 
     this.#sessionModule = new SessionModule(systemName, this.#logSystem);
+    this.#tokenModule = new TokenModule(systemName, this.#logSystem);
 
     this.#logSystem.debug(`End of ${systemName} creation`)
     this.#logSystem.info(`${systemName} initialization complete`);
@@ -45,6 +53,15 @@ export class StorageSystem extends SystemPlugin {
    */
   get session () {
     return this.#sessionModule;
+  }
+
+  /**
+   * Token module.
+   * @property @public
+   * @returns {TokenModule} TokenModule instance.
+   */
+  get tokenStorage() {
+    return this.#tokenModule;
   }
 
   /**