+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED]
+### Added
+- token module
+
 ### Changed
+- sdk version
 - plugin build mechanism
 - plugin inner architecture
   - remove `Vue` and `Vuex` dependencies